+++ Rules/IPv4/Outbound/WindowsSystem.ps1
@@ -671,7 +671,7 @@ if ((Test-ExecutableFile $Program) -or $ForceLoad)
 		-Service Any -Program $Program -Group $Group `
 		-Enabled True -Action Allow -Direction $Direction -Protocol TCP `
 		-LocalAddress Any -RemoteAddress Internet4 `
-		-LocalPort Any -RemotePort 443 `
+		-LocalPort Any -RemotePort 80, 443 `
 		-LocalUser $UsersGroupSDDL `
 		-InterfaceType $DefaultInterface `
 		-Description "Checks downloaded files, webpages and it analyzes pages and determines if