+++ docs/CHANGELOG.md
@@ -46,6 +46,7 @@ To see unreleased changes please do so on `develop` branch [HERE][changelog]\
   - Updated rules for store apps and pokerstars
   - Added rules for Motrix, yt-dlp and calibre
   - Removed invalid rules for vcpkg
+  - Updated rules for web browsers
 
 - Modules